requirements.txt
fastapi
uvicorn
sqlmodel
jinja2
python-multipart
yfinance
apscheduler
requests
python-dotenv